src/main/java/dmk/login/example/app/service/UserDetailsService.java
package dmk.login.example.app.service;

import dmk.login.example.app.dao.UserRepository;
import dmk.login.example.app.domain.User;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import static org.springframework.util.Assert.notNull;


@Service
public class UserDetailsService implements org.springframework.security.core.userdetails.UserDetailsService {

    private final UserRepository userRepository;

    @Autowired
    public UserDetailsService(UserRepository userRepository) {
        notNull(userRepository, "UserRepository can not be null");
        this.userRepository = userRepository;
    }

    @Override
    public UserDetails loadUserByUsername(String userName) throws UsernameNotFoundException {
        User user = userRepository.findByName(userName);
        if (user == null) {
            throw new UsernameNotFoundException(userName);
        }

        return new dmk.login.example.app.security.UserDetails(user);
    }
}
